+++ CHANGELOG.md
@@ -56,6 +56,8 @@ Line wrap the file at 100 chars.                                              Th
 - Fix OpenVPN warning about usage of AES-256-CBC cipher.
 - Fix "Out of time" screen status icon position.
 - Fix log newline characters on Windows.
+- Mullvad CLI can now be used with daemon instance that doesn't have the `--disable-rpc-auth`
+  flag set.
 
 
 ## [2018.1] - 2018-03-01